10-02-2025

+++ client/src/components/tablaReserves/TablaReserves.jsx
@@ -2,6 +2,7 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
 import "./tablaReserves.css";
+import { Alert } from "reactstrap";
 
 const TablaReserves = () => {
   const role = localStorage.getItem("role");
@@ -16,7 +17,9 @@ const TablaReserves = () => {
     }
   }, [role]);
 
+  // ****** RECUPERAR RESERVAS USUARIO ******
   const getReservesUser = async () => {
+
     try {
       const response = await axios.get("/api/myReserves", {
         withCredentials: true,
@@ -24,8 +27,9 @@ const TablaReserves = () => {
       if (response.data && response.data.success && response.data.reserves) {
         setReserves(response.data.reserves);
         setError("");
+        console.log("reservas " + reserves)
       } else {
-        setError("No hay reservas disponibles");
+        setError("No hay reservas solicitadas");
         setReserves([]);
       }
     } catch (error) {
@@ -35,6 +39,7 @@ const TablaReserves = () => {
     }
   };
 
+  // ****** RECUPERAR TODAS LAS RESERVAS ******
   const getReservesAdmin = async () => {
     try {
       const response = await axios.get("/api/reserves", {
@@ -43,41 +48,45 @@ const TablaReserves = () => {
       if (response.data && response.data.success && response.data.reserves) {
         setReserves(response.data.reserves);
         setError("");
+        console.log(reserves)
       } else {
-        setError("No hay reservas disponibles");
+        setError("No hay reservas solicitadas");
+        console.log("No hay reservas");
         setReserves([]);
       }
     } catch (error) {
       console.error("Error al obtener las reservas:", error);
+      console.log("Error al cargar las reservas");
       setError("Error al cargar las reservas");
       setReserves([]);
     }
   };
 
+  // ****** ELIMINAR RESERVA ******
   const handleDeleteReserve = async (reserveId) => {
-    try {
-      const response = await axios.delete(`/api/deleteReserve/${reserveId}`, {
-        withCredentials: true,
-      });
-      if (response.data && response.data.success) {
-        if (role == 0) {
-          getReservesUser();
-        } else if (role == 1) {
-          getReservesAdmin();
+    if (window.confirm('¿Estás seguro de que deseas eliminar esta reserva?')) {
+      try {
+        const response = await axios.delete(`/api/deleteReserve/${reserveId}`, {
+          withCredentials: true,
+        });
+        if (response.data && response.data.success) {
+          if (role == 0) {
+            getReservesUser();
+          } else if (role == 1) {
+            getReservesAdmin();
+          }
+        } else {
+          setError("Error al eliminar la reserva");
         }
-      } else {
+      } catch (error) {
+        console.error("Error al eliminar la reserva:", error);
         setError("Error al eliminar la reserva");
       }
-    } catch (error) {
-      console.error("Error al eliminar la reserva:", error);
-      setError("Error al eliminar la reserva");
     }
   };
 
   return (
     <div className="container tablaReserves col auto mt-4 mb-4">
-      {error && <div className="alert alert-danger">{error}</div>}
-
       <div className="reservesTitle text-center">
         <h1>RESERVAS</h1>
       </div>
@@ -91,41 +100,53 @@ const TablaReserves = () => {
                     <strong>Evento</strong>
                   </th>
                   <th scope="col">
+                    <strong>Fecha</strong>
+                  </th>
+                  <th scope="col">
                     <strong>Participante</strong>
                   </th>
                   <th scope="col">
-                    <strong>Acciones</strong>
+                    <strong></strong>
                   </th>
                 </tr>
               </thead>
               <tbody>
-                {reserves.map((reserve) => (
-                  <tr key={reserve._id}>
-                    <td>
-                      {reserve.event ? (
-                        reserve.event.name
-                      ) : (
-                        <span className="text-danger">
-                          Evento no disponible
-                        </span>
-                      )}
-                    </td>
-                    <td>
-                      {reserve.participating.name}{" "}
-                      {reserve.participating.surname}
-                    </td>
-                    <td>
-                      <button
-                        className="btn btn-danger btn-sm"
-                        onClick={() => handleDeleteReserve(reserve._id)}
-                      >
-                        Eliminar
-                      </button>
-                    </td>
-                  </tr>
-                ))}
+                {reserves
+                  .filter(reserve => reserve && reserve.event.dateActivity && new Date(reserve.event.dateActivity) >= new Date())
+                  .sort((a, b) => new Date(a.dateActivity) - new Date(b.dateActivity))
+                  .map((reserve) => (
+                    <tr key={reserve._id}>
+                      <td>
+                        {reserve.event ? (
+                          reserve.event.name
+                        ) : (
+                          <span className="text-danger">
+                            Evento no disponible
+                          </span>
+                        )}
+                      </td>
+                      <td>
+                        {reserve.event && reserve.event.dateActivity ?
+                          new Date(reserve.event.dateActivity).toLocaleString('es')
+                          : 'Fecha no disponible'
+                        }
+                      </td>
+                      <td>
+                        {reserve.participating.name}{" "}
+                        {reserve.participating.surname}
+                      </td>
+                      <td>
+                        <button
+                          className="btn btn-danger btn-sm"
+                          onClick={() => handleDeleteReserve(reserve._id)}
+                        >
+                          Eliminar
+                        </button>
+                      </td>
+                    </tr>
+                  ))}
               </tbody>
-            </table>
+            </table>{error && <div className="message">{error}</div>}
           </div>
         </div>
       </div>
